requirements.txt
asgiref==3.7.2
Cerberus==1.3.5
Django==4.2.6
djangorestframework==3.14.0
et-xmlfile==1.1.0
openpyxl==3.1.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
